pyproject.toml
[project]
name = "colorgen"
description = "Generate terminal colorschemes from images"
keywords = ["colorscheme", "theme", "terminal", "awesomewm", "kitty", "rofi", "waybar"]

version = "1.0.0"
requires-python = "==3.12.*"

license = "GPL-3.0-or-later"
readme = "README.md"

authors = [
    { name = 'Piotr Krzysztof', email = 'piotr@codextechnologies.org' }
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.12",
    "Operating System :: POSIX :: Linux",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Desktop Environment :: Window Managers",
]

homepage = "https://codeberg.org/piotrkrzysztof/colorgen"
repository = "https://codeberg.org/piotrkrzysztof/colorgen"

dependencies = [
    "pylette==5.1.1",
    "pillow<12.1.0",
]

[project.scripts]
colorgen = "colorgen:main"

[tool.poetry.dependencies]
python = "~3.12"
pylette = "==5.1.1"
pillow = "<12.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
ruff = "^0.14.12"
mypy = "^1.19.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.7.1"
mkdocstrings = {extras = ["python"], version = "^1.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=colour",
    "--cov=gen",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
branch = true
source = ["colour", "gen"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # Allow raising exceptions without from inside except
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # Allow assert in tests
    "ARG",    # Allow unused arguments in tests
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["pylette.*", "Pylette.*", "PIL.*"]
ignore_missing_imports = true
